requirements.txt
binaryornot==0.4.0
chardet==2.3.0
click==3.3
cookiecutter==1.0.0
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-allauth==0.23.0
django-crispy-forms==1.5.2
django-debug-toolbar==1.3.2
django-environ==0.4.0
django-extensions==1.5.7
django-toolbelt==0.0.1
djangorestframework==3.2.4
gunicorn==19.3.0
Jinja2==2.8
MarkupSafe==0.23
mock==1.3.0
oauthlib==1.0.3
pbr==1.8.0
psycopg2==2.6.1
python3-openid==3.0.7
PyYAML==3.11
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sqlparse==0.1.16
static3==0.6.1
wheel==0.24.0
